ポータルコンポーネントを作成
